use build script in Buildkite
