+++ .buildkite/pipeline.yml
@@ -2,23 +2,23 @@ steps:
 #  - label: "Format"
 #    command: make format
 #  - wait
-  - command: make api-test
-  - command: make reindex_worker-test
-  - command: make ingestor_works-test
-  - command: make ingestor_images-test
-  - command: make id_minter-test
-  - command: make matcher-test
-  - command: make merger-test
-  - command: make recorder-test
-  - command: make transformer_miro-test
-  - command: make transformer_sierra-test
-  - command: make transformer_mets-test
-  - command: make transformer_calm-test
-  - command: make sierra_reader-test
-  - command: make sierra_bib_merger-test
-  - command: make sierra_item_merger-test
-  - command: make sierra_items_to_dynamo-test
-  - command: make mets_adapter-test
-  - command: make calm_adapter-test
-  - command: make inference_manager-test
-  - command: make snapshot_generator-test
+  - command: .build/run_job.py api
+#  - command: make reindex_worker-test
+#  - command: make ingestor_works-test
+#  - command: make ingestor_images-test
+#  - command: make id_minter-test
+#  - command: make matcher-test
+#  - command: make merger-test
+#  - command: make recorder-test
+#  - command: make transformer_miro-test
+#  - command: make transformer_sierra-test
+#  - command: make transformer_mets-test
+#  - command: make transformer_calm-test
+#  - command: make sierra_reader-test
+#  - command: make sierra_bib_merger-test
+#  - command: make sierra_item_merger-test
+#  - command: make sierra_items_to_dynamo-test
+#  - command: make mets_adapter-test
+#  - command: make calm_adapter-test
+#  - command: make inference_manager-test
+#  - command: make snapshot_generator-test
